# Conflicts:
#	core/src/main/resources/core-services.xml
#	persistence/src/main/java/org/yes/cart/domain/entity/bridge/AttributeValueBridge.java
#	persistence/src/main/java/org/yes/cart/domain/entity/bridge/SkuWarehouseBridge.java
#	persistence/src/main/java/org/yes/cart/domain/entity/impl/ProductEntity.java
#	search/src/main/java/org/yes/cart/search/query/ProductSearchQueryBuilder.java
#	web/support/src/main/java/org/yes/cart/web/support/service/ProductServiceFacade.java
#	web/support/src/main/java/org/yes/cart/web/support/service/impl/ProductServiceFacadeImpl.java
#	web/support/src/main/java/org/yes/cart/web/support/util/ProductSortingUtils.java

+++ domain-api/src/main/java/org/yes/cart/search/dao/IndexBuilder.java
@@ -49,7 +49,7 @@ public interface IndexBuilder<T, PK extends Serializable> {
          *
          * @return count
          */
-        int getLastIndexCount();
+        long getLastIndexCount();
 
     }
